restructured files for create-grant
