--- app/src/pages/basics.tsx
@@ -1,20 +1,0 @@
-import type { NextPage } from "next";
-import Head from "next/head";
-import { BasicsView } from "../views";
-
-const Basics: NextPage = (props) => {
-  return (
-    <div>
-      <Head>
-        <title>Solana Scaffold</title>
-        <meta
-          name="description"
-          content="Basic Functionality"
-        />
-      </Head>
-      <BasicsView />
-    </div>
-  );
-};
-
-export default Basics;
